keras/requirements.txt
keras
libgpuarray
libprotobuf
mako
mock
pbr
protobuf
pygpu
tensorflow
theano
numpy
six
tensorflow
pyyaml
scipy
h5py
matplotlib